requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1
easyocr==1.7.1
Pillow==10.3.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
